requirements.txt
dhlab>=2.3.0
pandas==1.0.1
streamlit
requests==2.23.0
Pillow==7.0.0
urllib3==1.25.8
spacy>=3.3.0
spacy-transformers